requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-oidc-provider==0.7.0
future==0.17.1
gspread==3.1.0
httplib2==0.13.0
idna==2.8
line-bot-sdk==1.12.1
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodomex==3.8.2
pyjwkest==1.4.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
